requirements.txt
torch>=2.1.0
torchvision>=0.16.0
ttkbootstrap>=1.10.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0   # optional, for webcam capture
kaggle>=1.5.16         # optional, for dataset download
